+++ r/src/init.c
@@ -27,7 +27,7 @@ extern SEXP nanoarrow_c_is_altrep(SEXP x_sexp);
 extern SEXP nanoarrow_c_altrep_is_materialized(SEXP x_sexp);
 extern SEXP nanoarrow_c_altrep_force_materialize(SEXP x_sexp, SEXP recursive_sexp);
 extern SEXP nanoarrow_c_infer_ptype(SEXP array_xptr);
-extern SEXP nanoarrow_c_from_array(SEXP array_xptr, SEXP ptype_sexp);
+extern SEXP nanoarrow_c_materialize_array(SEXP array_xptr, SEXP ptype_sexp);
 extern SEXP nanoarrow_c_array_stream_get_schema(SEXP array_stream_xptr);
 extern SEXP nanoarrow_c_array_stream_get_next(SEXP array_stream_xptr);
 extern SEXP nanoarrow_c_array_view(SEXP array_xptr, SEXP schema_xptr);
@@ -60,7 +60,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"nanoarrow_c_altrep_is_materialized", (DL_FUNC)&nanoarrow_c_altrep_is_materialized, 1},
     {"nanoarrow_c_altrep_force_materialize", (DL_FUNC)&nanoarrow_c_altrep_force_materialize, 2},
     {"nanoarrow_c_infer_ptype", (DL_FUNC)&nanoarrow_c_infer_ptype, 1},
-    {"nanoarrow_c_from_array", (DL_FUNC)&nanoarrow_c_from_array, 2},
+    {"nanoarrow_c_materialize_array", (DL_FUNC)&nanoarrow_c_materialize_array, 2},
     {"nanoarrow_c_array_stream_get_schema", (DL_FUNC)&nanoarrow_c_array_stream_get_schema, 1},
     {"nanoarrow_c_array_stream_get_next", (DL_FUNC)&nanoarrow_c_array_stream_get_next, 1},
     {"nanoarrow_c_array_view", (DL_FUNC)&nanoarrow_c_array_view, 2},
