+++ r/src/materialize.h
@@ -96,15 +96,4 @@ int nanoarrow_materialize(struct ArrayViewSlice* src, struct VectorSlice* dst,
                           struct MaterializeOptions* options,
                           struct MaterializeContext* context);
 
-// These functions materialize a complete R vector or return R_NilValue
-// if they cannot (i.e., no conversion possible). These functions will warn
-// (once)  if there are values that cannot be converted (e.g., because they
-// are out of range).
-SEXP nanoarrow_materialize_unspecified(struct ArrowArrayView* array_view);
-SEXP nanoarrow_materialize_lgl(struct ArrowArrayView* array_view);
-SEXP nanoarrow_materialize_int(struct ArrowArrayView* array_view);
-SEXP nanoarrow_materialize_dbl(struct ArrowArrayView* array_view);
-SEXP nanoarrow_materialize_chr(struct ArrowArrayView* array_view);
-SEXP nanoarrow_materialize_list_of_raw(struct ArrowArrayView* array_view);
-
 #endif
